6.2 renamed types -> typings, added footerLogo for FavoritesScreen
