+++ src/components/map/map.tsx
@@ -1,8 +1,8 @@
 import {useRef, useEffect} from 'react';
 import {Icon, Marker, layerGroup} from 'leaflet';
 import useMap from '@hooks/use-map';
-import { City } from '../../types/city';
-import { Offer, Offers } from '../../types/offer';
+import { City } from '@typings/city';
+import { Offer, Offers } from '@typings/offer';
 import {URL_MARKER_DEFAULT, URL_MARKER_CURRENT} from '@const';
 import 'leaflet/dist/leaflet.css';
