+++ src/hooks/index.ts
@@ -1,5 +1,5 @@
 import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import type { State, AppDispatch } from '../types/state';
+import type { State, AppDispatch } from '@typings/state';
 
 export const useAppDispatch = () => useDispatch<AppDispatch>();
