+++ src/pages/favorites-screen/favorites-screen.tsx
@@ -1,9 +1,10 @@
 import Logo from '@components/logo/logo';
 import HeaderNav from '@components/header-nav/header-nav';
 import { Link } from 'react-router-dom';
-import { AppRoute, CardType } from '@const';
+import { CardType } from '@const';
 import PlaceCard from '@components/place-card/place-card';
 import { useAppSelector } from '@hooks/index';
+import FooterLogo from '@components/app/footer-logo/footer-logo';
 
 
 export default function FavoritesScreen(): JSX.Element {
@@ -62,11 +63,7 @@ export default function FavoritesScreen(): JSX.Element {
           </section>
         </div>
       </main>
-      <footer className="footer container">
-        <a className="footer__logo-link" href={AppRoute.Root}>
-          <img className="footer__logo" src="img/logo.svg" alt="6 cities logo" width="64" height="33" />
-        </a>
-      </footer>
+      <FooterLogo/>
     </div>
   );
 }
