+++ src/store/reducer.ts
@@ -1,12 +1,12 @@
 import { createReducer } from '@reduxjs/toolkit';
 import { setOffersList, changeCity, setReviews, setOffersInDetails, setSortType } from './action';
-import { Offers } from '../types/offer';
-import { Reviews } from '../types/review';
-import { OffersInDetails } from '../types/offerInDetails';
+import { Offers } from '@typings/offer';
+import { Reviews } from '@typings/review';
+import { OffersInDetails } from '@typings/offerInDetails';
+import { City } from '@typings/city';
 import { offers } from '@mocks/offers';
 import { reviews } from '@mocks/reviews';
 import { offersInDetails } from '@mocks/offersInDetails';
-import { City } from '../types/city';
 import { Cities, SortType } from '@const';
