+++ src/typings/offer.ts
