+++ src/typings/offerInDetails.ts
